+++ sagemaker-core/src/sagemaker/core/__init__.py
@@ -1,3 +1,4 @@
+"""SageMaker Core package for low-level resource management and SDK foundations."""
 from sagemaker.core.utils.utils import enable_textual_rich_console_and_traceback