+++ sagemaker-core/src/sagemaker/core/remote_function/invoke_function.py
@@ -98,7 +98,7 @@ def _load_pipeline_context(args) -> Context:
 
 
 def _execute_remote_function(
-    sagemaker_session, s3_base_uri, s3_kms_key, run_in_context, context
+    sagemaker_session, s3_base_uri, s3_kms_key, run_in_context, signing_key, context
 ):
     """Execute stored remote function"""
     from sagemaker.core.remote_function.core.stored_function import StoredFunction
@@ -107,6 +107,7 @@ def _execute_remote_function(
         sagemaker_session=sagemaker_session,
         s3_base_uri=s3_base_uri,
         s3_kms_key=s3_kms_key,
+        signing_key=signing_key,
         context=context,
     )
 
@@ -137,12 +138,15 @@ def main(sys_args=None):
         run_in_context = args.run_in_context
         pipeline_context = _load_pipeline_context(args)
 
+        signing_key = os.getenv("REMOTE_FUNCTION_SECRET_KEY")
+
         sagemaker_session = _get_sagemaker_session(region)
         _execute_remote_function(
             sagemaker_session=sagemaker_session,
             s3_base_uri=s3_base_uri,
             s3_kms_key=s3_kms_key,
             run_in_context=run_in_context,
+            signing_key=signing_key,
             context=pipeline_context,
         )