+++ sagemaker-mlops/tests/unit/workflow/test_pipeline.py
@@ -361,6 +361,7 @@ def test_get_function_step_result_incomplete_job(mock_session):
         "AlgorithmSpecification": {"ContainerEntrypoint": JOBS_CONTAINER_ENTRYPOINT},
         "OutputDataConfig": {"S3OutputPath": "s3://bucket/path"},
         "TrainingJobStatus": "Failed",
+        "Environment": {"REMOTE_FUNCTION_SECRET_KEY": "key"}
     }
     
     with pytest.raises(RemoteFunctionError, match="not in Completed status"):
@@ -376,7 +377,7 @@ def test_get_function_step_result_success(mock_session):
         "AlgorithmSpecification": {"ContainerEntrypoint": JOBS_CONTAINER_ENTRYPOINT},
         "OutputDataConfig": {"S3OutputPath": "s3://bucket/path/exec-id/step1/results"},
         "TrainingJobStatus": "Completed",
-        "Environment": {},
+        "Environment": {"REMOTE_FUNCTION_SECRET_KEY": "key"}
     }
     
     with patch("sagemaker.mlops.workflow.pipeline.deserialize_obj_from_s3", return_value="result"):
@@ -499,7 +500,7 @@ def test_pipeline_execution_result_terminal_failure(mock_session):
         "AlgorithmSpecification": {"ContainerEntrypoint": JOBS_CONTAINER_ENTRYPOINT},
         "OutputDataConfig": {"S3OutputPath": "s3://bucket/path/exec-id/step1/results"},
         "TrainingJobStatus": "Completed",
-        "Environment": {},
+        "Environment": {"REMOTE_FUNCTION_SECRET_KEY": "key"}
     }
     
     with patch.object(execution, "wait", side_effect=WaiterError("name", "Waiter encountered a terminal failure state", {})):
@@ -517,7 +518,7 @@ def test_get_function_step_result_obsolete_s3_path(mock_session):
         "AlgorithmSpecification": {"ContainerEntrypoint": JOBS_CONTAINER_ENTRYPOINT},
         "OutputDataConfig": {"S3OutputPath": "s3://bucket/different/path"},
         "TrainingJobStatus": "Completed",
-        "Environment": {},
+        "Environment": {"REMOTE_FUNCTION_SECRET_KEY": "key"}
     }
     
     with patch("sagemaker.mlops.workflow.pipeline.deserialize_obj_from_s3", return_value="result"):